feat: Add Create Route and Create Service functionalities with UI and controller logic
- Implemented CreateRouteView for route creation with form fields for origin, start date, end date, odometer readings, and file attachment.
- Added CreateServiceView for service creation with fields for date, time, odometer, service details, and file attachment.
- Created CreateRouteController and CreateServiceController to handle business logic and data management for route and service creation.
- Integrated image picker functionality for file attachments in both views.
- Established bindings for CreateRoute and CreateService to manage dependencies.
- Enhanced validation for form fields to ensure data integrity.

+++ lib/custom-widget/report/charts/monthly_cost_bar_chart.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
 class MonthlyCostBarChart extends StatelessWidget {
-  final Map<String, double> monthlyData;
+  final Map<String, int> monthlyData;
   final bool isUrdu;
   final String title;
 
@@ -110,7 +110,7 @@ class MonthlyCostBarChart extends StatelessWidget {
   double _getMaxY() {
     double max = 0;
     for (var value in monthlyData.values) {
-      if (value > max) max = value;
+      if (value > max) max = value.toDouble();
     }
     return max * 1.2;
   }
@@ -159,7 +159,7 @@ class MonthlyCostBarChart extends StatelessWidget {
         x: index,
         barRods: [
           BarChartRodData(
-            toY: values[index],
+            toY: values[index].toDouble(),
             color: Utils.appColor,
             width: 16,
             borderRadius: const BorderRadius.only(
